Modified TreeFileHandler. Added separate file for constants. Fixed tree handler test.

+++ tests/test_tree_file_handler.py
@@ -3,11 +3,9 @@ from unittest import TestCase
 
 from rtree.node import Node
 from rtree.tree_file_handler import TreeFileHandler
+from rtree.default_config import *
 
-TREE_FILE_TEST = "testingTree.bin"
 
-
-#
 # def test_tree_file_handler_write(): # move to separate file
 #     os.remove(TREE_FILE_TEST)
 #     tree_file_handler = TreeFileHandler(TREE_FILE_TEST, 2, 1024)
@@ -35,41 +33,48 @@ TREE_FILE_TEST = "testingTree.bin"
 
 class TestTreeFileHandler(TestCase):
     def test_write_read_node_one_handler_instance(self):
-        os.remove(TREE_FILE_TEST)
+        print("TEST test_write_read_node_one_handler_instance: ") if PRINT_OUTPUT_TEST else None
+        if os.path.isfile(TREE_FILE_TEST):
+            os.remove(TREE_FILE_TEST)
+
         tree_file_handler_writer = TreeFileHandler(TREE_FILE_TEST, 2, 1024)
         written_node_1 = Node(False, [[1, 1], [2, 2]], [5, 5, 2, 2, 2])
         written_node_2 = Node(True, [[5, 4], [3, 2]], [1, 2])
         id1 = tree_file_handler_writer.write_node(written_node_1)
         id2 = tree_file_handler_writer.write_node(written_node_2)
-        tree_file_handler_writer.__del__()
-
-        tree_file_handler_writer = TreeFileHandler(TREE_FILE_TEST, 2, 1024)
 
         read_node1 = tree_file_handler_writer.get_node(id1)
         read_node2 = tree_file_handler_writer.get_node(id2)
 
-        print(written_node_1)
-        print(read_node1)
+        print(written_node_1) if PRINT_OUTPUT_TEST else None
+        print(read_node1) if PRINT_OUTPUT_TEST else None
+        print(written_node_2) if PRINT_OUTPUT_TEST else None
+        print(read_node2) if PRINT_OUTPUT_TEST else None
 
         assert written_node_1.__dict__ == read_node1.__dict__
         assert written_node_2.__dict__ == read_node2.__dict__
 
     def test_write_read_node_two_handler_instances(self):
-        os.remove(TREE_FILE_TEST)
+        print("TEST test_write_read_node_two_handler_instances") if PRINT_OUTPUT_TEST else None
+        if os.path.isfile(TREE_FILE_TEST):
+            os.remove(TREE_FILE_TEST)
+
         tree_file_handler_writer = TreeFileHandler(TREE_FILE_TEST, 2, 1024)
         written_node_1 = Node(False, [[1, 1], [2, 2]], [])
         written_node_2 = Node(True, [[5, 4], [3, 2]], [1, 2])
         id1 = tree_file_handler_writer.write_node(written_node_1)
         id2 = tree_file_handler_writer.write_node(written_node_2)
-        tree_file_handler_writer.__del__()
 
+        tree_file_handler_writer.__del__()
         tree_file_handler_reader = TreeFileHandler(TREE_FILE_TEST, 2, 1024)
 
         read_node1 = tree_file_handler_reader.get_node(id1)
         read_node2 = tree_file_handler_reader.get_node(id2)
 
-        print(written_node_1)
-        print(read_node1)
+        print(written_node_1) if PRINT_OUTPUT_TEST else None
+        print(read_node1) if PRINT_OUTPUT_TEST else None
+        print(written_node_2) if PRINT_OUTPUT_TEST else None
+        print(read_node2) if PRINT_OUTPUT_TEST else None
 
         assert written_node_1.__dict__ == read_node1.__dict__
         assert written_node_2.__dict__ == read_node2.__dict__
